# public/openapi.yaml
openapi: 3.1.0
info:
  title: Lune API
  version: 0.3.0
  description: >
    API publique (auth, santé) et API app (PERSONAL / PRO). Toutes les IDs BigInt sont sérialisées en string.
    Toutes les réponses mutables exigent CSRF (Origin) et renvoient `x-request-id` + `cache-control: no-store`.
servers:
  - url: http://localhost:3000
    description: Développement local
  - url: https://diwanbg.work
    description: Production (Railway)
security:
  - cookieAuth: []
...
paths:
...
  /api/pro/businesses/{businessId}/services:
    get:
      tags: [Pro / Services]
      security: [ { cookieAuth: [] } ]
      summary: Liste des services catalogue
      parameters:
        - $ref: '#/components/parameters/BusinessId'
        - in: query
          name: q
          required: false
          schema: { type: string }
        - in: query
          name: type
          required: false
          schema: { type: string }
        - in: query
          name: categoryReferenceId
          required: false
          schema: { type: string }
        - in: query
          name: tagReferenceId
          required: false
          schema: { type: string }
      responses:
        '200': { $ref: '#/components/responses/Ok' }
    post:
      tags: [Pro / Services]
      security: [ { cookieAuth: [] } ]
      summary: Crée un service catalogue
      parameters: [ { $ref: '#/components/parameters/BusinessId' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, code]
              properties:
                code: { type: string }
                name: { type: string }
                type: { type: string }
                defaultPriceCents: { type: integer }
                categoryReferenceId: { type: string, nullable: true }
                tagReferenceIds:
                  type: array
                  items: { type: string }
                description: { type: string, nullable: true }
                tjmCents: { type: integer, nullable: true }
                durationHours: { type: integer, nullable: true }
                vatRate: { type: integer, nullable: true }
      responses:
        '201': { $ref: '#/components/responses/Ok' }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: auth_token
  parameters:
    BusinessId:
      name: businessId
      in: path
      required: true
      schema: { type: string }
...
  responses:
...
    Ok:
      description: Opération réussie
      content:
        application/json:
          schema:
            type: object
            properties:
              ok: { type: boolean, example: true }
...