# src/app/api/pro/businesses/[businessId]/services/route.ts
import { NextRequest, NextResponse } from 'next/server';
import { prisma } from '@/server/db/client';
import { requireAuthPro } from '@/server/auth/requireAuthPro';
import { requireBusinessRole } from '@/server/auth/businessRole';
import { assertSameOrigin, jsonNoStore, withNoStore } from '@/server/security/csrf';
import { rateLimit } from '@/server/security/rateLimit';
import { badRequest, getRequestId, unauthorized, withRequestId } from '@/server/http/apiUtils';
import { BusinessReferenceType, TaskPhase } from '@/generated/prisma/client';

function forbidden() {
  return NextResponse.json({ error: 'Forbidden' }, { status: 403 });
}

function withIdNoStore(res: NextResponse, requestId: string) {
  return withNoStore(withRequestId(res, requestId));
}

function parseId(param: string | undefined) {
  if (!param || !/^\d+$/.test(param)) return null;
  try {
    return BigInt(param);
  } catch {
    return null;
  }
}

function normalizeStr(v: unknown) {
  return String(v ?? '').trim();
}

function isRecord(v: unknown): v is Record<string, unknown> {
  return !!v && typeof v === 'object';
}

function validateCode(code: string) {
  if (!code) return 'Code requis.';
  if (!/^SER-[0-9A-Za-z_-]+$/.test(code)) return 'Code invalide (format SER-XXX).';
  if (code.length > 50) return 'Code trop long.';
  return null;
}

type ServiceTemplateInput = {
  phase: TaskPhase | null;
  title: string;
  defaultAssigneeRole: string | null;
  defaultDueOffsetDays: number | null;
};

type ServiceBodyParsed =
  | { error: string }
  | {
      code: string;
      name: string;
      categoryReferenceId: bigint | null | undefined;
      tagReferenceIds: bigint[] | undefined;
      type: string | null;
      description: string | null;
      defaultPriceCents: number | null;
      tjmCents: number | null;
      durationHours: number | null;
      vatRate: number | null;
      templates: ServiceTemplateInput[];
    };

function validateServiceBody(body: unknown): ServiceBodyParsed {
  if (!isRecord(body)) return { error: 'Payload invalide.' };
  const code = normalizeStr(body.code);
  const name = normalizeStr(body.name);
  const type = normalizeStr(body.type);
  const description = normalizeStr(body.description);
  const defaultPriceCents =
    typeof body.defaultPriceCents === 'number' && Number.isFinite(body.defaultPriceCents)
      ? Math.max(0, Math.trunc(body.defaultPriceCents))
      : null;
  const tjmCents =
    typeof body.tjmCents === 'number' && Number.isFinite(body.tjmCents)
      ? Math.max(0, Math.trunc(body.tjmCents))
      : null;
  const durationHours =
    typeof body.durationHours === 'number' && Number.isFinite(body.durationHours)
      ? Math.max(0, Math.trunc(body.durationHours))
      : null;
  const vatRate =
    typeof body.vatRate === 'number' && Number.isFinite(body.vatRate)
      ? Math.max(0, Math.trunc(body.vatRate))
      : null;

  const codeErr = validateCode(code);
  if (codeErr) return { error: codeErr };
  if (!name) return { error: 'Nom requis.' };
  if (name.length > 140) return { error: 'Nom trop long (140 max).' };
  if (description && description.length > 2000) return { error: 'Description trop longue.' };

  const templates: ServiceTemplateInput[] = Array.isArray(body.taskTemplates)
    ? body.taskTemplates.filter(isRecord).map((t) => {
        const phaseRaw = typeof t.phase === 'string' ? t.phase : null;
        const phase =
          phaseRaw && Object.values(TaskPhase).includes(phaseRaw as TaskPhase)
            ? (phaseRaw as TaskPhase)
            : null;
        return {
          phase,
          title: normalizeStr(t.title),
          defaultAssigneeRole: normalizeStr(t.defaultAssigneeRole || '') || null,
          defaultDueOffsetDays:
            typeof t.defaultDueOffsetDays === 'number' && Number.isFinite(t.defaultDueOffsetDays)
              ? Math.trunc(t.defaultDueOffsetDays)
              : null,
        };
      })
    : [];

  for (const tpl of templates) {
    if (!tpl.title) return { error: 'Template tâche : titre requis.' };
    if (tpl.title.length > 180) return { error: 'Template tâche : titre trop long.' };
    if (tpl.defaultAssigneeRole && tpl.defaultAssigneeRole.length > 120)
      return { error: 'Template tâche : assigneeRole trop long.' };
  }

  const categoryProvided = Object.prototype.hasOwnProperty.call(body, 'categoryReferenceId');
  const categoryReferenceId =
    categoryProvided && typeof body.categoryReferenceId === 'string' && /^\d+$/.test(body.categoryReferenceId)
      ? BigInt(body.categoryReferenceId)
      : categoryProvided
        ? null
        : undefined;

  const tagProvided = Object.prototype.hasOwnProperty.call(body, 'tagReferenceIds');
  const tagReferenceIds: bigint[] | undefined = tagProvided
    ? Array.from(
        new Set(
          (Array.isArray(body.tagReferenceIds) ? body.tagReferenceIds : [])
            .filter((id) => typeof id === 'string' && /^\d+$/.test(id))
            .map((id) => BigInt(id as string))
        )
      )
    : undefined;

  return {
    code,
    name,
    categoryReferenceId,
    tagReferenceIds,
    type: type || null,
    description: description || null,
    defaultPriceCents,
    tjmCents,
    durationHours,
    vatRate,
    templates,
  };
}

function ensureServiceDelegate(requestId: string) {
  if (!(prisma as { service?: unknown }).service) {
    return withIdNoStore(
      NextResponse.json(
        { error: 'Prisma client not generated / wrong import (service delegate absent).' },
        { status: 500 }
      ),
      requestId
    );
  }
  return null;
}

async function validateCategoryAndTags(
  businessId: bigint,
  categoryReferenceId: bigint | null,
  tagReferenceIds?: bigint[]
): Promise<{ categoryId: bigint | null; tagIds: bigint[] } | { error: string }> {
  if (categoryReferenceId) {
    const category = await prisma.businessReference.findFirst({
      where: {
        id: categoryReferenceId,
        businessId,
        type: BusinessReferenceType.CATEGORY,
        isArchived: false,
      },
      select: { id: true },
    });
    if (!category) {
      return { error: 'categoryReferenceId invalide pour ce business.' };
    }
  }

  let tagIds: bigint[] = [];
  if (tagReferenceIds && tagReferenceIds.length) {
    const tags = await prisma.businessReference.findMany({
      where: {
        id: { in: tagReferenceIds },
        businessId,
        type: BusinessReferenceType.TAG,
        isArchived: false,
      },
      select: { id: true },
    });
    if (tags.length !== tagReferenceIds.length) {
      return { error: 'tagReferenceIds invalides pour ce business.' };
    }
    tagIds = tags.map((t) => t.id);
  }

  return { categoryId: categoryReferenceId, tagIds };
}

// GET /api/pro/businesses/{businessId}/services
export async function GET(
  request: NextRequest,
  context: { params: Promise<{ businessId: string }> }
) {
  const requestId = getRequestId(request);
  let userId: string;
  try {
    ({ userId } = await requireAuthPro(request));
  } catch {
    return withIdNoStore(unauthorized(), requestId);
  }
  const { businessId } = await context.params;
  const businessIdBigInt = parseId(businessId);
  if (!businessIdBigInt) return withIdNoStore(badRequest('businessId invalide.'), requestId);

  const delegateError = ensureServiceDelegate(requestId);
  if (delegateError) return delegateError;

  const membership = await requireBusinessRole(businessIdBigInt, BigInt(userId), 'VIEWER');
  if (!membership) return withIdNoStore(forbidden(), requestId);

  const { searchParams } = new URL(request.url);
  const q = searchParams.get('q')?.trim();
  const type = searchParams.get('type')?.trim();
  const categoryReferenceIdParam = searchParams.get('categoryReferenceId');
  const tagReferenceIdParam = searchParams.get('tagReferenceId');

  const categoryReferenceId = categoryReferenceIdParam ? parseId(categoryReferenceIdParam) : null;
  if (categoryReferenceIdParam && !categoryReferenceId) {
    return withIdNoStore(badRequest('categoryReferenceId invalide.'), requestId);
  }
  const tagReferenceId = tagReferenceIdParam ? parseId(tagReferenceIdParam) : null;
  if (tagReferenceIdParam && !tagReferenceId) {
    return withIdNoStore(badRequest('tagReferenceId invalide.'), requestId);
  }

  const services = await prisma.service.findMany({
    where: {
      businessId: businessIdBigInt,
      ...(q
        ? {
            OR: [
              { name: { contains: q, mode: 'insensitive' } },
              { code: { contains: q, mode: 'insensitive' } },
              { description: { contains: q, mode: 'insensitive' } },
            ],
          }
        : {}),
      ...(type ? { type: { equals: type, mode: 'insensitive' } } : {}),
      ...(categoryReferenceId ? { categoryReferenceId } : {}),
      ...(tagReferenceId ? { tags: { some: { referenceId: tagReferenceId } } } : {}),
    },
    orderBy: [{ createdAt: 'desc' }],
    include: {
      _count: { select: { taskTemplates: true } },
      categoryReference: { select: { id: true, name: true } },
      tags: { include: { reference: { select: { id: true, name: true } } } },
    },
  });

  return withIdNoStore(
    jsonNoStore({
      items: services.map((s) => ({
        id: s.id.toString(),
        businessId: s.businessId.toString(),
        code: s.code,
        name: s.name,
        type: s.type,
        description: s.description,
        categoryReferenceId: s.categoryReferenceId ? s.categoryReferenceId.toString() : null,
        categoryReferenceName: s.categoryReference?.name ?? null,
        tagReferences: s.tags.map((t) => ({ id: t.reference.id.toString(), name: t.reference.name })),
        defaultPriceCents: s.defaultPriceCents?.toString() ?? null,
        tjmCents: s.tjmCents?.toString() ?? null,
        durationHours: s.durationHours,
        vatRate: s.vatRate,
        templateCount: s._count?.taskTemplates ?? 0,
        createdAt: s.createdAt.toISOString(),
        updatedAt: s.updatedAt.toISOString(),
      })),
    }),
    requestId
  );
}

// POST /api/pro/businesses/{businessId}/services
export async function POST(
  request: NextRequest,
  context: { params: Promise<{ businessId: string }> }
) {
  const requestId = getRequestId(request);
  const csrf = assertSameOrigin(request);
  if (csrf) return withIdNoStore(csrf, requestId);

  let userId: string;
  try {
    ({ userId } = await requireAuthPro(request));
  } catch {
    return withIdNoStore(unauthorized(), requestId);
  }

  const { businessId } = await context.params;
  const businessIdBigInt = parseId(businessId);
  if (!businessIdBigInt) return withIdNoStore(badRequest('businessId invalide.'), requestId);

  const membership = await requireBusinessRole(businessIdBigInt, BigInt(userId), 'ADMIN');
  if (!membership) return withIdNoStore(forbidden(), requestId);

  const delegateError = ensureServiceDelegate(requestId);
  if (delegateError) return delegateError;

  const limited = rateLimit(request, {
    key: `pro:services:create:${businessIdBigInt}:${userId}`,
    limit: 120,
    windowMs: 60 * 60 * 1000,
  });
  if (limited) return withIdNoStore(limited, requestId);

  const body = await request.json().catch(() => null);
  const parsed = validateServiceBody(body);
  if ('error' in parsed) return withIdNoStore(badRequest(parsed.error), requestId);

  const validated = await validateCategoryAndTags(
    businessIdBigInt,
    parsed.categoryReferenceId ?? null,
    parsed.tagReferenceIds
  );
  if ('error' in validated) {
    return withIdNoStore(badRequest(validated.error), requestId);
  }

  try {
    const created = await prisma.service.create({
      data: {
        businessId: businessIdBigInt,
        code: parsed.code,
        name: parsed.name,
        categoryReferenceId: validated.categoryId ?? undefined,
        type: parsed.type || undefined,
        description: parsed.description || undefined,
        defaultPriceCents: parsed.defaultPriceCents ?? undefined,
        tjmCents: parsed.tjmCents ?? undefined,
        durationHours: parsed.durationHours ?? undefined,
        vatRate: parsed.vatRate ?? undefined,
        tags:
          validated.tagIds.length > 0
            ? {
                create: validated.tagIds.map((id) => ({ referenceId: id })),
              }
            : undefined,
        taskTemplates: parsed.templates.length
          ? {
              create: parsed.templates.map((tpl) => ({
                phase: tpl.phase ?? undefined,
                title: tpl.title,
                defaultAssigneeRole: tpl.defaultAssigneeRole || undefined,
                defaultDueOffsetDays: tpl.defaultDueOffsetDays ?? undefined,
              })),
            }
          : undefined,
      },
      include: { taskTemplates: true },
    });

    return withIdNoStore(
      NextResponse.json(
        {
          id: created.id.toString(),
          businessId: created.businessId.toString(),
          code: created.code,
          name: created.name,
          categoryReferenceId: created.categoryReferenceId ? created.categoryReferenceId.toString() : null,
          type: created.type,
          description: created.description,
          defaultPriceCents: created.defaultPriceCents?.toString() ?? null,
          tjmCents: created.tjmCents?.toString() ?? null,
          durationHours: created.durationHours,
          vatRate: created.vatRate,
          templateCount: created.taskTemplates.length,
          tagReferences: validated.tagIds.map((id) => ({ id: id.toString() })),
          taskTemplates: created.taskTemplates.map((tpl) => ({
            id: tpl.id.toString(),
            phase: tpl.phase,
            title: tpl.title,
            defaultAssigneeRole: tpl.defaultAssigneeRole,
            defaultDueOffsetDays: tpl.defaultDueOffsetDays,
          })),
          createdAt: created.createdAt.toISOString(),
          updatedAt: created.updatedAt.toISOString(),
        },
        { status: 201 }
      ),
      requestId
    );
  } catch (err) {
    console.error(err);
    return withIdNoStore(badRequest('Création impossible.'), requestId);
  }
}